Set web.first_key only on the first uploadStart message
UploadStart is sent by the phone app whenever a data line is appended
to an empty queue. It can be useful to compute the time spent sending
data, compared to the time spent waiting for more data to be sent from
the watch. However only the furst uploadStart message is useful for
determining the first uploaded key.